base1.1

+++ BOT-Bacbo/bot_bacbo_base.py
@@ -232,7 +232,7 @@ class BOT_Bacbo:
                     elif i == 'Tie':
                         results.append("E")
 
-                if check != results:
+                if check != results and len(results) > 5:
                     check = results
                     self.delete()
                     self.estrategy(results)
