+++ BOT-FootballStudio/bot_footballstudio_base.py
@@ -233,7 +233,7 @@ class BOT_FootballStudio:
                     elif i == 'Tie':
                         results.append("E")
 
-                if check != results:
+                if check != results and len(results) > 5:
                     check = results
                     self.delete()
                     self.estrategy(results)
