+++ BOT-Roulette/bot_roleta_base.py
@@ -274,7 +274,7 @@ class BOT_Roleta:
                 for i in json_data:
                     results.append(int(i))
 
-                if check != results:
+                if check != results and len(results) > 5:
                     check = results
                     self.delete()
                     self.estrategy(results)
